feat: integrate Cycles context provider and refactor Home and History components to use context
